Add files via upload
This is a .html file that uses Python code to duplicate most of the analyses and visualization's from the RNotebook deliverable file. For more information on the research question and data that this code was used to investigate, see this publication: https://moralitylab.bc.edu/wp-content/uploads/2022/01/mcmanus2021final.pdf.